GH-35: [feat] Insure routing includes profile page after rebasing
